re-organize the code (#26)
